Bump Conda env PyTorch version to 1.13.*
Because of
```
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
torchvision 0.14.1+cu116 requires torch==1.13.1, but you have torch 1.12.1 which is incompatible.
torchtext 0.14.1 requires torch==1.13.1, but you have torch 1.12.1 which is incompatible.
torchaudio 0.13.1+cu116 requires torch==1.13.1, but you have torch 1.12.1 which is incompatible.
```